pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "gendiff"
authors = ["ysemenuk <invent@list.ru>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.8"
flake8 = "^5.0.4"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
main = "gendiff.scripts.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
